separating settings pagees into separate files, button text realignment
