+++ frontend/Goals.jsx
@@ -1,0 +1,17 @@
+import * as React from 'react';
+import {
+  Text,
+} from 'react-native';
+import {
+  BoilerplateView,
+} from './Themes';
+
+function Goals() {
+  return (
+    <BoilerplateView>
+      <Text>Boilerplate for Goals</Text>
+    </BoilerplateView>
+  );
+}
+
+export default Goals;
