+++ frontend/Notifications.jsx
@@ -1,0 +1,17 @@
+import * as React from 'react';
+import {
+  Text,
+} from 'react-native';
+import {
+  BoilerplateView,
+} from './Themes';
+
+function Notifications() {
+  return (
+    <BoilerplateView>
+      <Text>Boilerplate for Notifications</Text>
+    </BoilerplateView>
+  );
+}
+
+export default Notifications;
