+++ frontend/Preferences.jsx
@@ -1,0 +1,17 @@
+import * as React from 'react';
+import {
+  Text,
+} from 'react-native';
+import {
+  BoilerplateView,
+} from './Themes';
+
+function Preferences() {
+  return (
+    <BoilerplateView>
+      <Text>Boilerplate for Preferences</Text>
+    </BoilerplateView>
+  );
+}
+
+export default Preferences;
