+++ frontend/Profile.jsx
@@ -1,0 +1,17 @@
+import * as React from 'react';
+import {
+  Text,
+} from 'react-native';
+import {
+  BoilerplateView,
+} from './Themes';
+
+function Profile() {
+  return (
+    <BoilerplateView>
+      <Text>Boilerplate for Profile</Text>
+    </BoilerplateView>
+  );
+}
+
+export default Profile;
